[jit] Delete the ErrorReport default constructor
An error report without a SourceRange context is bad, because it doesn't
tell the user where something happened. Delete the default constructor
to make it harder to create errors like this (you can still use a fake
SourceRange if you absolutely need to).

Also clean up the only case where the default constructor was used.

ghstack-source-id: 7f1c912bd45dfb7a281675429b5a277f77f7b85d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32879