Correctly preserve "not set anywhere" TensorOptions when merging.
Previously, we would fill in TensorOption with defaults whenever an
item was missing from both the left and right side of the merge.  This
is morally incorrect: if we don't have an item on the left or right,
we should keep the entry empty (so the downstream user can apply
the appropriate defaulting rule).

I don't think this caused any bugs, but I noticed this error when
working on a later patch in my diff stack.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 6e2d3ab8b7493045439d4051a657688d02e8c9c8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33510