Stop generating out full function type for registration, use decltype or infer it
Previously, we had to specify full types because the functions we registering
might be overloaded, and the type was necessary to resolve the ambiguity.  I
disambiguate all of these names by mangling the names of the methods we
place on CPUType/CUDAType/TypeDefault with the overload name (these are
*internal* wrappers which are not user visible), and then can strip
the generation of full function types from the registration.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 07ff5ddcc92ecc3f72fc86dc5a7092e92dc60e39
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33097